requirements.txt
numpy>=1.16.5,<1.23.0
click>=8.0
glfw==2.2.0
pyopengl==3.1.5
imageio-ffmpeg==0.4.3
tensorboard==2.4.1
pillow==8.3.1
scipy==1.7.2
requests==2.26.0
ninja==1.10.2
matplotlib==3.4.2
imageio==2.9.0
opencv-python==4.5.5.64 
av==8.1.0
torch_tb_profiler==0.3.1
joblib==1.1.0
protobuf==3.20.0
trimesh
imgui
pyspng
scikit-image
tqdm
moviepy
plyfile
plotly
wandb
gitpython
hydra-core
rich
clean-fid
lpips